requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
jinja2
# Authentication
python-jose
passlib
python-multipart
httpx
msal
itsdangerous

# AI/LLM
langchain
langchain-anthropic
langchain-google-genai

# PDF Processing
PyPDF2

# Utilities
pydantic